requirements.txt
asgiref==3.3.1
bleach==3.3.0
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
dynaconf==3.1.3
idna==2.10
importlib-metadata==3.7.0
Markdown==3.3.4
mimesis==4.1.3
packaging==20.9
pyparsing==2.4.7
pytz==2021.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
webencodings==0.5.1
zipp==3.4.0
